refactor(table): extract update map handling into its own module
Dataset config, table metadata, schema metadata and per-field metadata are all
string maps updated the same way: entries where a None value deletes the key,
plus a flag choosing between merge and replace. The type, its four From
conversions and the three functions that apply it move together.
